Rename alert-table-header.tsx => AlertTableHeader.tsx
