+++ extensions/ql-vscode/src/view/results/AlertTableHeader.tsx
